\name{myres}
\alias{myres}
\title{Dummy function to prepare for model2ssfun.
}
\description{
   Dummy function.
}
\usage{
   myres(prm, ...)
}
\arguments{
  \item{prm}{
     A parameter vector
}
  \item{\dots}{
     Any data needed for computation of the residual vector from the expression
       rhsexpression - lhsvar. Note that this is the negative of the usual residual,
     but the sum of squares is the same.
}
}
\details{
     none.
}
\value{
     none
}
\author{
John C Nash <nashjc@uottawa.ca>
}

